+++ TabbedPanelMod/history.md
@@ -1,5 +1,11 @@
 # History
 
+## v0.2
+
+* Added screen shot image
+* Modified the logic on how/when the widths of the panels get stored
+* other minor mods
+
 ## v0.1
 
 * Release version